pyproject.toml
[project]
name = "polybot-nextprice"
version = "0.1.0"
description = "Next-price / short-horizon prediction for prediction markets (Kalshi/Polymarket)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "duckdb>=0.9",
    "scikit-learn>=1.3",
    "torch>=2.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
